ragbot/LocalAIAgentWithRAG-main/requirements.txt
# Core LangChain & AI
langchain>=0.0.350
langchain-community>=0.0.50
langchain-ollama
langchain-core
langchain-text-splitters

# Document Processing
pypdf>=4.0.0
python-docx>=1.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Scientific Computing (for vector store)
numpy>=1.21.0